refactor(registry): Move packument fetching into separate module
